riscv64: Fix isel for widening and narrowing operations
The riscv64 isel cannot rely on always extending a 32/16/8-bit integer
result to 64 bits and then utilizing this fact to treat Iop_<N>Sto64 as
a no-op. The issue is that helpers implemented by the tools can return
a value that is not sign-extended and therefore breaks the described
premise.