riscv64: Add correct name for the runtime linker
This change fixes a problem with Helgrind reporting errors also from
ld.so. These can be valid errors but are not useful for the user and so
should be normally hidden by the tool. In particular, this silences
a race on _rtld_local._dl_num_relocations from _dl_lookup_symbol_x()
which was reported in many tests and made them fail.